f32/f16 -> fp32/fp16

+++ models/turbine_models/tests/sd_test.py
@@ -35,7 +35,7 @@ default_arguments = {
     "batch_size": 1,
     "height": 512,
     "width": 512,
-    "precision": "fp16",
+    "precision": "fp32",
     "max_length": 77,
     "guidance_scale": 7.5,
     "run_vmfb": True,
